extracted tiles to their own files, cast ints to floats rect blueprint tile painter render

+++ BlueprintTilePainter.cpp
@@ -1,6 +1,7 @@
 #include "BlueprintTilePainter.h"
 #include <SDL3/SDL.h>
 #include "Vectors.h"
+#include "Tile.h"
 
 BlueprintTilePainter::BlueprintTilePainter(SpaceShipBlueprint* blueprint, ShipBuildingGrid* grid, Tile tileToPaint) : Entity(Vector2Int(0, 0), std::nullopt), blueprint(blueprint), grid(grid), tileToPaint(tileToPaint)
 {
@@ -52,7 +53,12 @@ void BlueprintTilePainter::render(SDL_Renderer* renderer, const RenderingContext
 
 	Vector2Float toPaintPos = context.toScreenPoint(coords);
 
-	SDL_FRect rect = { toPaintPos.x,toPaintPos.y,pxGridSize / context.cameraScale,pxGridSize / context.cameraScale };
+	SDL_FRect rect = { 
+		static_cast<float>(toPaintPos.x),
+		static_cast<float>(toPaintPos.y),
+		static_cast<float>(pxGridSize / context.cameraScale),
+		static_cast<float>(pxGridSize / context.cameraScale)
+	};
 	SDL_SetRenderDrawColor(renderer,0,0,255,255);
 	SDL_RenderRect(renderer, &rect);
 }
